+++ tests/test_tools.py
@@ -159,8 +159,7 @@ def test_openai_tools():
 
 @skip_if_no_anthropic
 def test_anthropic_tools():
-    check_tools("anthropic/claude-3-sonnet-20240229", tools_beta=False)
-    check_tools("anthropic/claude-3-sonnet-20240229", tools_beta=True)
+    check_tools("anthropic/claude-3-sonnet-20240229")
 
 
 @skip_if_no_mistral